// src/interfaces/references/OrderItemFeature.ts
export interface OrderItemFeature {
  stockCode: string;
  itemCode: string;
  feature1Type: string | null;
  feature1Name: string | null;
  feature2Type: string | null;
  feature2Name: string | null;
  feature3Type: string | null;
  feature3Name: string | null;
  feature4Type: string | null;
  feature4Name: string | null;
  costPerUnit: number | null;
}